requirements.txt
oslo.config
pyyaml
